refactor : 에러코드 수정 및 OpenAIClient -> ChatGPTClient 클래스명 변경, HashMap->객체사용
