+++ src/main/java/com/fifteen/auction/global/client/chatgpt/dto/ChatMessage.java
@@ -1,0 +1,11 @@
+package com.fifteen.auction.global.client.chatgpt.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ChatMessage {
+    private String role;
+    private String content;
+}
